Enhance guest addition for share
This PR introduces the following changes:
* the search in the guest popover now also proposes Cozy url based on
what is present in the contact application;
* the calendar application can now handle a contact for whom several
Cozy instances are linked;
* if a guest has a Cozy url and an email in his contact information then
the user can decide if she wants to share the invitation or send it by
mail. It is possible to switch between those two possibilities thanks to
two new buttons. The currently selected option has its button
highlighted in blue;
* if it is not possible to switch from email to share, or from share to
email the button corresponding to the unavailable choice is hidden;
* a Cozy url can be typed directly in the search field. If no "@" is
found it is considered as the url of a Cozy. If a "@" is found it is
interpreted as an email address;
* if an email address or a Cozy url is too long to be displayed inside
the popover then the overflowing text is replaced by three dots and is
only displayed on two lines when hovering.